Rename result4j-assertj to assertj-result4j, fixes for review comments.
