+++ .gitignore
@@ -1,4 +1,4 @@
 /target
 /result4j/target
-/result4j-assertj/target
+/assertj-result4j/target
 *~
